pyproject.toml
[project]
name = "meal-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "notion-client>=3.0.0",
    "protobuf>=7.34.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
    "google-genai>=1.10.0",
]
